requirements.txt
numpy>=1.26,<3.0
opencv-contrib-python>=4.10,<5.0
PyYAML>=6.0,<7.0
airsim>=1.8.1
pytest>=8.0,<9.0
